% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspm1d-init.R
\docType{class}
\name{bspm1dData-class}
\alias{bspm1dData-class}
\title{bspm1dData}
\description{
An S4 class bspm1dData
}
\section{Slots}{

\describe{
\item{\code{data}}{An object of class data.frame containing the data in long format.}
}}

