% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspm1d-init.R
\name{bspm1dData}
\alias{bspm1dData}
\title{bspm1dData}
\usage{
bspm1dData(data, dimname, outcome, id, group = NA_character_)
}
\arguments{
\item{data}{....}

\item{dimname}{....}

\item{outcome}{....}

\item{id}{....}

\item{group}{....}
}
\description{
bspm1dData
}
