% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspm-bfttest.R
\name{bspm_bf2pp}
\alias{bspm_bf2pp}
\title{bspm_bf2pp}
\usage{
bspm_bf2pp(bf)
}
\arguments{
\item{bf}{an object of class \code{BFBayesFactor}}
}
\description{
Transform the Bayes Factor to a posterior probability, assuming equal prior odds for the competing hypotheses. See BayesFactor::newPriorOdds() and BayesFactor::as.BFprobability().
}
