% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspm-bfttest.R
\name{bspm_bfpp}
\alias{bspm_bfpp}
\title{bspm_bfpp}
\usage{
bspm_bfpp(data, group, dimension, outcome, paired, nullInterval, rscale)
}
\description{
Pointwise calculation of the BayesFactor and transformation to posterior probabilities.
}
\seealso{
\code{\link[=bspm_bfttest(), bspm_bf(), bspm_bf2pp]{bspm_bfttest(), bspm_bf(), bspm_bf2pp()}}
}
