% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspm-bfttest.R
\name{bspm_bfttest}
\alias{bspm_bfttest}
\title{bspm_bfttest}
\usage{
bspm_bfttest(
  data,
  group,
  dimension,
  outcome,
  paired,
  nullInterval = c(-0.2, 0.2),
  rscale = "medium"
)
}
\arguments{
\item{data}{data.frame in long format containing the dataset}

\item{group}{name of the grouping variable (paired or independent groups) (column in \code{data})}

\item{dimension}{name of the variable with the 1-dimensional domain (column in \code{data})}

\item{outcome}{name of the outcome variable (column in \code{data})}

\item{paired}{logical, are the data from a paired (TRUE) or independent group (FALSE) design?, see BayesFactor::ttestBF()}

\item{nullInterval}{see BayesFactor::ttestBF()}

\item{rscale}{see BayesFactor::ttestBF()}
}
\value{
description
}
\description{
This function calculates a Bayes-Factor version of the SPM{t}-test for 1-dimensional data (two-sample or paired sample t-test) as described in Serrien et al. (2019).
}
\references{
Serrien, B., Goossens, M., & Baeyens, J. P. (2019). Statistical parametric mapping of biomechanical one-dimensional data with Bayesian inference. International Biomechanics, vol. 6 (1), 9-18
}
