% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bspm_plot1d}
\alias{bspm_plot1d}
\title{bspm_plot1d}
\usage{
bspm_plot1d(
  data,
  outcome = NULL,
  dimension = NULL,
  grp_factors = NULL,
  err = "sd",
  level = 0.95
)
}
\arguments{
\item{data}{R data.frame containing the individual-level 1-dimensional data in long format.}

\item{outcome}{String, specifying the name of the outcome variable (required).}

\item{dimension}{String, specifyng the name of the 1-dimensional domain (required).}

\item{grp_factors}{String, specifying the name of grouping factors in the data (optional).}

\item{err}{String, specifying the type of error-cloud to calculate (required). Should be either \code{sd} or \code{ci}. In case of \code{ci}, a level of confidence should be defined (default .95).}

\item{level}{Numeric, level of confidence for the confidence interval (default .95).}
}
\value{
A ggplot figure is returned.
}
\description{
This function calculates a mean +/- error cloud for the 1-dimensional dataset. The error cloud can be either based on the SD or a confidence interval. Grouping factors can be specified.
}
