% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspm1d-summarise.R
\name{bspm_summarise}
\alias{bspm_summarise}
\title{bspm_summarise}
\usage{
bspm_summarise(object, err = "sd", level = 0.95)
}
\arguments{
\item{object}{An S4 object of class bspm1dData, result from a call to bspm_data().}

\item{err}{String, specifying the type of error-cloud to calculate (required). Should be either \code{sd}, \code{se} or \code{ci}. In case of \code{ci}, a level of confidence should be defined.}

\item{level}{Numeric, level of confidence for the confidence interval (default .95).}
}
\value{
Tibble containing summary statistics of the outcome variable at each point in the 1-dimensional domain (stratified per grouping factor). For an outcome variable called \code{Y}, the summary statistics are called \code{Y_m} (mean), \code{Y_sd} (standard deviation of the data), \code{Y_se} (standard error of the mean), \verb{Y_cil, Y_ciu} (lower and upper limits of the confidence interval).
}
\description{
This function calculates a mean +/- error cloud for the 1-dimensional dataset.
}
