% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspm1d-summarise.R
\name{bspm_summarise1d}
\alias{bspm_summarise1d}
\title{setClass(
"bspm1dSummary",
slots = c(
summary = "data.frame"
),
prototype = list(
summary = NULL
)
)
bspm_summarise1d}
\usage{
bspm_summarise1d(
  data,
  outcome = NULL,
  dimension = NULL,
  grp_factors = NULL,
  err = "sd",
  level = 0.95
)
}
\arguments{
\item{data}{R data.frame containing the individual-level 1-dimensional data in long format.}

\item{outcome}{String, specifying the name of the outcome variable (required).}

\item{dimension}{String, specifyng the name of the 1-dimensional domain (required).}

\item{grp_factors}{String, specifying the name of grouping factors in the data (optional).}

\item{err}{String, specifying the type of error-cloud to calculate (required). Should be either \code{sd}, \code{se} or \code{ci}. In case of \code{ci}, a level of confidence should be defined (default .95).}

\item{level}{Numeric, level of confidence for the confidence interval (default .95).}
}
\value{
Tibble containing summary statistics of the outcome variable at each point in the 1-dimensional domain (stratified per grouping factor). For an outcome variable called \code{Y}, the summary statistics are called \code{Y_m} (mean), \code{Y_sd} (standard deviation of the data), \code{Y_se} (standard error of the mean), \verb{Y_cil, Y_ciu} (lower and upper limits of the confidence interval).
}
\description{
This function calculates a mean +/- error cloud for the 1-dimensional dataset. The error cloud can be either based on the SD or a confidence interval. Grouping factors can be specified to calculate group-specific summaries.
}
