% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.bspm_build_aes}
\alias{.bspm_build_aes}
\title{.bspm_build_aes}
\usage{
.bspm_build_aes(outcome, dimension, grp_factors)
}
\arguments{
\item{outcome}{String, name of the outcome variable.}

\item{dimension}{String, name of the x-axis variable.}

\item{grp_factors}{String or NULL, name of the grouping variable.}
}
\value{
An aes() object.
}
\description{
Internal helper: builds the ggplot aes(), optionally adding color/fill
when a grouping factor is supplied.
}
