% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspm1d-init.R
\name{bspm1dData}
\alias{bspm1dData}
\title{bspm1dData}
\usage{
bspm1dData(data, dimension, outcome, id, paired = FALSE, group = NA_character_)
}
\arguments{
\item{data}{data.frame that contains the 1-dimensional data; the outcome vector mist be a list-column where each row contains a 1xD row vector (D elements in the 1-dimensional domain)}

\item{dimension}{1-dimensional domain in which the outcome is measured, can be a simple vector like 0:100 or can contain units like units::set_units(0:100, "\%")}

\item{outcome}{character with name of the outcome variable}

\item{id}{ID-variable, must be able to identify paired data}

\item{paired}{logical, must be TRUE when 2 paired observations per \code{id} are present (1 in each level of \code{group})}

\item{group}{optional vector that defines a grouping factor (2-levels)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
