% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspm-bfttest.R
\name{bspm_bfttest}
\alias{bspm_bfttest}
\title{bspm_bfttest}
\usage{
bspm_bfttest(data)
}
\arguments{
\item{data}{}
}
\description{
This function calculates a Bayes-Factor version of the t-test (two-sample or paired sample t-test).
}
